feat: Add support for user-defined tags
Introduces a new `tags` table in the database to store user-defined tags.
This allows users to create custom tags for their entries, enhancing data
categorization and analysis capabilities. The `TagEntry` type and associated
methods for managing tags have been added to the `DataContext` and `StorageService`.
The database schema has been updated to version 5 to accommodate this new table.